requirements.txt
pyyaml
requests
html5lib
flask
lxml
flask-bootstrap4
Flask-Cache
sqlalchemy-utils
bs4
feedparser